+++ CHANGELOG.md
@@ -12,5 +12,12 @@ and here
 
 ### Changed
 -->
+## Release 2.1.5 -- 2024-03
+- added engine.startEvent
+### cli enhancements
+- support double quotes, so parameters like: `name "Buy Used Car" ` can now work
+- support backspace on input
+
+
 ## Release 2.1.0 -- 2024-03
-added engine.restart
+- added engine.restart